change size field to big integer

+++ django_project/gap/models/dataset.py
@@ -174,7 +174,7 @@ class DataSourceFileCache(models.Model):
         null=True,
         blank=True
     )
-    size = models.PositiveIntegerField(default=0)
+    size = models.PositiveBigIntegerField(default=0)
 
     class Meta:
         """Meta class for DataSourceFileCache."""
